refactor: starbucksStorage.js 폴더 이동
